requirements.txt
Flask==3.0.3
SQLAlchemy==2.0.30
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
Flask-Login==0.6.3
Werkzeug==3.0.3